Conflicts:
	exp/dataset.py
	exp/demo.py

+++ exp/run_liblinear.py
@@ -29,48 +29,51 @@ def run(dataset, model_config, output_file):
         result_file   = dst_folder + '/result_%g.txt' %(c)
         test_file   = dst_folder + '/test_%g.txt' %(c)
 
-        #clear the result file
-        open(test_file,'w').close()
-
         result_once = util.ResultItem()
+        if os.path.exists(result_file) == True:
+            result_once.load_result(result_file)
+        else:
+            #clear the result file
+            open(test_file,'w').close()
 
-        #training 
-        train_cmd = train_exe + ' -s 5 -c %f' %c + ' %s' %dataset.get_train_file(model_config['rand_num']) + ' %s' %model_file
-        train_cmd = train_cmd.replace('/',os.sep)
-        print train_cmd
-        start_time =time.time()
-        os.system(train_cmd)
-        end_time = time.time()
 
-        #hard to evaluate train_error, set to zero
-        result_once.append_value('train_error',0)
+            #training 
+            train_cmd = train_exe + ' -s 5 -c %f' %c + ' %s' %dataset.get_train_file(model_config['rand_num']) + ' %s' %model_file
+            train_cmd = train_cmd.replace('/',os.sep)
+            print train_cmd
+            start_time =time.time()
+            os.system(train_cmd)
+            end_time = time.time()
 
-        #parse learning time
-        train_time = (float)(end_time - start_time) 
-        result_once.append_value('train_time',train_time)
+            #hard to evaluate train_error, set to zero
+            result_once.append_value('train_error',0)
 
-        #predict
-        test_cmd = test_exe + ' %s' %dataset.test_file + ' %s' %model_file + ' %s' %predict_file + '>> %s' %test_file
-        test_cmd = test_cmd.replace('/',os.sep)
-        print test_cmd
-        start_time =time.time()
-        os.system(test_cmd)
-        end_time = time.time()
-        test_time = (float)(end_time - start_time) 
+            #parse learning time
+            train_time = (float)(end_time - start_time) 
+            result_once.append_value('train_time',train_time)
 
-        result_once.append_value('test_time',test_time)
+            #predict
+            test_cmd = test_exe + ' %s' %dataset.test_file + ' %s' %model_file + ' %s' %predict_file + '>> %s' %test_file
+            test_cmd = test_cmd.replace('/',os.sep)
+            print test_cmd
+            start_time =time.time()
+            os.system(test_cmd)
+            end_time = time.time()
+            test_time = (float)(end_time - start_time) 
 
-        test_error = parse_test_error_rate(test_file)
-        result_once.append_value('test_error',test_error)
+            result_once.append_value('test_time',test_time)
 
-        model_size = get_model_size(model_file)
+            test_error = parse_test_error_rate(test_file)
+            result_once.append_value('test_error',test_error)
 
-        result_once.append_value('non_zero_num',model_size)
+            model_size = get_model_size(model_file)
 
-        sparse_rate = 100.0 - model_size * 100.0 / dataset.dim
-        result_once.append_value('sparse_rate', sparse_rate)
+            result_once.append_value('non_zero_num',model_size)
 
-        result_once.save_result(result_file)
+            sparse_rate = 100.0 - model_size * 100.0 / dataset.dim
+            result_once.append_value('sparse_rate', sparse_rate)
+
+            result_once.save_result(result_file)
 
         result.Append(result_once)
